add elixir to contribute pr docs, add purl tests, create some refactoring
