+++ .github/workflows/semantic-pr.yaml
@@ -70,6 +70,7 @@ jobs:
             go
             c
             c++
+            dart
 
             os
             lang
