+++ pkg/fanal/analyzer/all/import.go
@@ -6,7 +6,7 @@ import (
 	_ "github.com/aquasecurity/trivy/pkg/fanal/analyzer/config/all"
 	_ "github.com/aquasecurity/trivy/pkg/fanal/analyzer/executable"
 	_ "github.com/aquasecurity/trivy/pkg/fanal/analyzer/language/c/conan"
-	_ "github.com/aquasecurity/trivy/pkg/fanal/analyzer/language/dart/pubspeclock"
+	_ "github.com/aquasecurity/trivy/pkg/fanal/analyzer/language/dart/pubspec"
 	_ "github.com/aquasecurity/trivy/pkg/fanal/analyzer/language/dotnet/deps"
 	_ "github.com/aquasecurity/trivy/pkg/fanal/analyzer/language/dotnet/nuget"
 	_ "github.com/aquasecurity/trivy/pkg/fanal/analyzer/language/golang/binary"
