+++ pkg/fanal/analyzer/language/dart/pubspec/testdata/happy.lock
